requirements.txt
python-telegram-bot==20.7
requests==2.31.0
clarifai==2.6.2
Pillow==10.1.0
python-dotenv==1.0.0

